requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.2
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
langchain==0.0.340
langchain-core==0.1.0
langchain-google-genai==0.0.1
google-generativeai==0.3.0
openai==1.3.6
transformers==4.35.0
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.7.2
sqlalchemy==2.0.21
pydantic==2.5.0
beautifulsoup4==4.12.2
lxml==4.9.3
flask==3.0.0
flask-cors==4.0.0
matplotlib
numpy
pandas
pillow
plotly
pydantic
python-dotenv
requests
scikit-learn
seaborn
streamlit
google-genai
